ML-Interview-Preparation/MlOps-Project/End-to-End-Mlops-Image-Classification-Project/CI-CD-Pipeline/amazon-sagemaker-pipelines-serverless-inference/container/docker/requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
torchvision==0.12.0+cpu
alibi-detect[torch]
pillow
boto3